Refactor audit into a mixin, rename, properly filter.
