+++ packages/core/src/app.ts
@@ -4,7 +4,7 @@ import { Serializable, Serialize, AsDate, AsSerializable, bytesToBase64, stringT
 import { Invite, InvitePurpose } from "./invite";
 import { Vault, VaultID } from "./vault";
 import { Org, OrgID, OrgMember, OrgRole, Group, UnlockedOrg, OrgInfo } from "./org";
-import { VaultItem, VaultItemID, Field, Tag, createVaultItem } from "./item";
+import { VaultItem, VaultItemID, Field, Tag, createVaultItem, AuditResult } from "./item";
 import { Account, AccountID, UnlockedAccount } from "./account";
 import { Auth } from "./auth";
 import { Session, SessionID } from "./session";
@@ -1440,6 +1440,8 @@ export class App {
             fields?: Field[];
             tags?: Tag[];
             attachments?: AttachmentInfo[];
+            auditResults?: AuditResult[];
+            lastAudited?: Date;
         }
     ) {
         const { vault } = this.getItem(item.id)!;
